+++ src/components/statusBullet/index.js
@@ -1,3 +1,0 @@
-import StatusBullet from './StatusBullet';
-
-export default StatusBullet;