+++ src/index.js
@@ -26,7 +26,7 @@ import Checkbox from './components/checkbox';
 import Container from './components/container';
 import Counter from './components/counter';
 import DataGrid from './components/datagrid';
-import { DatePicker, DatePickerRange, DatePickerInput, DatePickerInputRange } from './components/datepicker';
+import { DatePicker, DatePickerInput } from './components/datepicker';
 import DetailPage, { DetailPageBody, DetailPageHeader } from './components/detailPage';
 import Dialog, { DialogBase } from './components/dialog';
 import EmptyState from './components/emptyState';
@@ -58,12 +58,10 @@ import Section from './components/section';
 import Select, { AsyncSelect } from './components/select';
 import { Island, IslandGroup } from './components/island';
 import SplitButton from './components/splitButton';
-import StatusBullet from './components/statusBullet';
 import StatusLabel from './components/statusLabel';
 import { TitleTab, TabGroup } from './components/tab';
 import Tag from './components/tag';
 import Timer from './components/timer';
-import TimerPulser from './components/timerPulser';
 import { Toast, ToastContainer } from './components/toast';
 import Toggle from './components/toggle';
 import ValidationText, { ErrorText, HelpText } from './components/validationText';
@@ -106,8 +104,6 @@ export {
   DataGrid,
   DatePicker,
   DatePickerInput,
-  DatePickerRange,
-  DatePickerInputRange,
   DetailPage,
   DetailPageBody,
   DetailPageHeader,
@@ -165,7 +161,6 @@ export {
   Section,
   Select,
   SplitButton,
-  StatusBullet,
   StatusLabel,
   TextBody,
   TextBodyCompact,
@@ -177,7 +172,6 @@ export {
   TimeInput,
   DurationInput,
   Timer,
-  TimerPulser,
   TitleTab,
   Toast,
   ToastContainer,